Made the basic separation

+++ app/controllers/battle_levels_controller.rb
@@ -4,6 +4,7 @@ class BattleLevelsController < ApplicationController
 
   def create
     @battle_level = BattleLevel.new battle_level_params
+    authorize @battle_level
     if @battle_level.save
       redirect_to parties_index_path, notice: "Battle Level Created"
     else
@@ -12,6 +13,7 @@ class BattleLevelsController < ApplicationController
   end
 
   def destroy
+    authorize @battle_level
     if @battle_level.destroy
       redirect_to parties_index_path, notice: "Battle Level Removed"
     else
@@ -23,6 +25,7 @@ class BattleLevelsController < ApplicationController
   end
 
   def update
+    authorize @battle_level
     @battle_level.update_attributes(battle_level_params)
     if @battle_level.save
       redirect_to parties_index_path, notice: "Success!"
